requirements.txt
# Core Framework
pydantic>=2.5.0
pydantic-ai>=0.0.24

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0

# Azure Integration
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pgvector>=0.2.4

# Observability
langfuse>=2.27.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
structlog>=24.1.0

# Document Processing
pypdf>=4.0.0
pdf2image>=1.17.0
pytesseract>=0.3.10
python-docx>=1.1.0

# Data Generation & Testing
faker>=22.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
hypothesis>=6.98.0
httpx>=0.26.0

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1
tenacity>=8.2.3

# Terminal Visualization
rich>=13.7.0
pyfiglet>=1.0.2

# MCP Server (Model Context Protocol)
mcp>=1.0.0
